requirements.txt
streamlit
jupyter
ipykernel
langchain
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
langchain-text-splitters
faiss-cpu
mistralai
pymilvus
pdfplumber
